pyproject.toml
[project]
name = "claw-ea"
version = "0.1.0"
description = "MCP server for medical office automation — Obsidian, Apple Calendar, Reminders"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "pyyaml>=6.0",
    "pyobjc-framework-EventKit>=10.0",
    "pyobjc-framework-Vision>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claw_ea"]

[tool.pytest.ini_options]
markers = [
    "macos: tests requiring macOS APIs (EventKit, Vision)",
    "converter: tests requiring real converter CLIs (docling, markitdown)",
]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
